Add Show 3D button
