+++ index.css
@@ -7,9 +7,10 @@
 @import url("./view/layout/index.css");
 @import url("./view/sample/index.css");
 @import url("./view/title/index.css");
-@import url("./view/banner/index.css");
+@import url("./view/link/index.css");
 @import url("./view/mode/index.css");
 @import url("./view/code/index.css");
+@import url("./view/button/index.css");
 @import url("./view/checkbox/index.css");
 @import url("./view/settings/index.css");
 @import url("./view/model/index.css");
