+++ index.ts
@@ -9,4 +9,5 @@ import './view/mode/index.js'
 import './view/code/index.js'
 import './view/checkbox/index.js'
 import './view/settings/index.js'
+import './view/model/index.js'
 import './view/benchmark/index.js'
