+++ stores/mode.ts
@@ -1,10 +1,7 @@
 import { atom } from 'nanostores'
 
-import { showP3, showRec2020 } from './settings.js'
-
 export let isBenchmark = atom(false)
 export let is3d = atom(false)
-export let loading3D = atom(false)
 
 if (/[?&]bench(=|&|$|&)/.test(location.search)) {
   isBenchmark.set(true)
@@ -24,28 +21,3 @@ function update(): void {
 
 isBenchmark.listen(update)
 is3d.listen(update)
-
-async function init(): Promise<void> {
-  loading3D.set(true)
-  let { generate3d } = await import('../view/model/index.js')
-  loading3D.set(false)
-
-  generate3d(showP3.get(), showRec2020.get())
-  showP3.listen(() => {
-    generate3d(showP3.get(), showRec2020.get())
-  })
-  showRec2020.listen(() => {
-    generate3d(showP3.get(), showRec2020.get())
-  })
-}
-
-if (is3d.get()) {
-  init()
-} else {
-  let unbindLoad = is3d.listen(async value => {
-    if (value) {
-      unbindLoad()
-      init()
-    }
-  })
-}
