+++ view/button/index.css
@@ -1,0 +1,14 @@
+.button {
+  height: 28px;
+  padding: 0 16px;
+  font-size: 10px;
+  color: var(--accent);
+  background: var(--surface-ui-accent);
+  border: none;
+  border-radius: 14px;
+
+  &:hover {
+    color: var(--text-badge);
+    background: var(--accent);
+  }
+}
