+++ view/button/mixin.pug
@@ -1,0 +1,2 @@
+mixin button(text, id)
+  button.button( data-button=id )= text
