+++ view/model/index.css
@@ -1,5 +1,19 @@
 .model {
-  &.is-hidden {
-    display: none;
+  display: flex;
+  justify-content: center;
+  height: 0;
+  overflow: hidden;
+  background: var(--surface-2);
+
+  :is(body.was-loaded) & {
+    transition: height 400ms;
   }
+
+  &.is-shown {
+    height: 500px;
+  }
+}
+
+.model_canvas {
+  height: 500px;
 }
