+++ view/model/index.pug
@@ -1,3 +1,3 @@
-.model.is-hidden( aria-hidden="true" )
+.model( aria-hidden="true" )
   .model_loader( aria-hidden="true" )
   canvas.model_canvas
