+++ view/title/index.css
@@ -2,8 +2,5 @@
   padding: 28px 0;
   font-size: 17.5px;
   line-height: 20px;
-
-  @media (width < 830px) {
-    text-align: center;
-  }
+  text-align: center;
 }
